requirements.txt
numpy==1.23.3
opencv-contrib-python-headless
pandas==1.5.1
Pillow==9.2.0
plotly==5.11.0
plotly-express==0.4.1
protobuf==3.20.1
requests==2.28.1
deta
python-dotenv